feat(mcp-local-server-wikitoolkit): TASK-2076 — Extract MCPToolAdapter + MCPResource to core
