+++ packages/ai-parrot-server/src/parrot/mcp/resources.py
@@ -1,26 +1,4 @@
-from dataclasses import dataclass
-from typing import Optional, Any
+"""Shim — MCPResource moved to core in FEAT-403."""
+from parrot.mcp.resources import MCPResource
 
-@dataclass
-class MCPResource:
-    """
-    Represents an MCP Resource.
-    
-    Resources are read-only data sources exposed by the server.
-    """
-    uri: str
-    name: str
-    description: Optional[str] = None
-    mime_type: Optional[str] = None
-    
-    def to_dict(self) -> dict[str, Any]:
-        """Convert to MCP protocol dictionary."""
-        data = {
-            "uri": self.uri,
-            "name": self.name,
-        }
-        if self.description:
-            data["description"] = self.description
-        if self.mime_type:
-            data["mimeType"] = self.mime_type
-        return data
+__all__ = ["MCPResource"]
